Extract mixer code to mixer.rs
